pyproject.toml
[tool.poetry]
name = "repl_nix_career-website"
version = "0.1.0"
description = ""
authors = ["Replit user <<>>"]

[tool.poetry.dependencies]
python = "^3.10"
Flask = "^2.3.1"
Flask-SQLAlchemy = "^3.1.1"
requests = "^2.32.3"
scikit-learn = "^1.5.1"
openai = "^1.57.0"
pandas = "^2.2.3"
python-dotenv = "^1.0.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
